Adds the test automation framework
This java maven project automates tests for the given website using Selenium webdriver and Cucumber/BDD.
It's implemented through Page Object Model.
There is a Cucumber Hook class, which has required methods to be executed before and after scenarios/steps.
It takes Screenshot for the failed scenarios which is getting saved is Output/Failed_Screenshot
It publishes Cucumber report after the test completes.